+++ server/apps/core/services/site_services.py
@@ -4,10 +4,10 @@ from typing import Optional
 
 from django.db import transaction
 
-from apps.core.models import HaztrakSite
 from apps.core.services import RcrainfoService, get_rcrainfo_client
 from apps.manifest.services import EManifest, PullManifestsResult, TaskResponse
 from apps.manifest.tasks import sync_site_manifests
+from apps.site.models import HaztrakSite
 
 logger = logging.getLogger(__name__)
