+++ server/apps/org/admin.py
@@ -1,8 +1,9 @@
 from django.contrib import admin
 
-from apps.core.models import HaztrakProfile, HaztrakSite
+from apps.core.models import HaztrakProfile
 from apps.org.models import HaztrakOrg
 from apps.rcrasite.admin import HaztrakSiteInline
+from apps.site.models import HaztrakSite
 
 
 class HaztrakProfileInline(admin.TabularInline):
